Add Numberer class to GeoUtils.IO.
Numberer solves the following problem: When optimizing for load
balancing, the patches must be numbered in a certain way according to
their sizes.  This numbering must be accounted for when writing the
topology sets in the IFEM input file.  We don't want to (a) hardcode a
specific numbering/load balancing in each mesh generation script or (b)
carry around an index map when specifying topology sets.

A typical way to use Numberer is as follows:

- Instantiate a Numberer and add all the patches
- Group the patches into (potentially overlapping) groups
- Use the groups to define topology sets
- Call Renumber() with the desired number of CPUs
- Call WriteEverything() to output both G2 file and IFEM input